Debug election

+++ node/src/io/RandomNumberGenerator.hpp
@@ -11,6 +11,6 @@ public:
 private:
   std::random_device seed{};
   std::mt19937_64 engine{seed()};
-  std::uniform_int_distribution<std::uint64_t> generator{1, UINT64_MAX};
+  std::uniform_int_distribution<std::uint64_t> generator{1, UINT16_MAX};
 };
 } // namespace io
